eg-etl-using-crm/requirements.txt
# Requires Python >=3.8,<4.0
Flask>=3.0.0,<4.0.0
requests>=2.31.0,<3.0.0
pandas>=2.2.1,<3.0.0
psycopg2-binary>=2.9.9,<3.0.0
